Refactor table resize tests to use React Testing Library and TypeScript
